// src/app/dashboardportal/procurement/sr/createSR/types/srTypes.ts
export type SupplierBranchOption = {
	party_mst_branch_id: number;
	state_id: number | null;
	state_name: string;
	gst_no?: string | null;
};
// ...
export type SRHeader = {
	inward_id: number;
	inward_no: string;
	inward_date: string;
	branch_id: number;
	branch_name: string;
	supplier_id: number;
	supplier_name: string;
	supplier_branch_id: number | null;
	supplier_branch_options: SupplierBranchOption[];
	supplier_state_id: number | null;
	supplier_state_name: string;
	bill_branch_id: number | null;
	billing_branch_name: string;
	billing_state_id: number | null;
	billing_state_name: string;
	ship_branch_id: number | null;
	shipping_branch_name: string;
	shipping_state_id: number | null;
	shipping_state_name: string;
	india_gst: boolean;
	sr_no: string;
	sr_date: string;
	sr_status: number;
	sr_status_name: string;
	inspection_date: string;
	challan_no: string;
	challan_date: string;
	invoice_no: string;
	invoice_date: string;
	invoice_amount: number;
	invoice_recvd_date: string;
	vehicle_number: string;
	driver_name: string;
	driver_contact_no: string;
	consignment_no: string;
	consignment_date: string;
	ewaybillno: string;
	ewaybill_date: string;
	despatch_remarks: string;
	receipts_remarks: string;
	sr_remarks: string;
	gross_amount: number;
	net_amount: number;
};